requirements.txt
pip
Flask
requests
json2html
cassandra-driver
flask-cassandra
